+++ exporters/otlp/src/otlp_grpc_exporter_factory.cc
@@ -27,6 +27,15 @@ std::unique_ptr<opentelemetry::sdk::trace::SpanExporter> OtlpGrpcExporterFactory
   return exporter;
 }
 
+std::unique_ptr<opentelemetry::sdk::trace::SpanExporter> OtlpGrpcExporterFactory::Create(
+    const OtlpGrpcExporterOptions &options,
+    nostd::shared_ptr<OtlpGrpcClient> client)
+{
+  std::unique_ptr<opentelemetry::sdk::trace::SpanExporter> exporter(
+      new OtlpGrpcExporter(options, std::move(client)));
+  return exporter;
+}
+
 }  // namespace otlp
 }  // namespace exporter
 OPENTELEMETRY_END_NAMESPACE